requirements.txt
python-dotenv==1.1.1
polars==1.31.0
sentence-transformers==4.1.0
gigachat==0.1.39.post2
numpy==2.3.1
scikit-learn==1.7.0